Fix handling CallbackBinding when key is contained in JSON fields
As previously the CallbackBindings where moved into a separate property in the ClassBindings class, they need to be called independently of the keys being present in the JSON fields, because the hasBindings method doesn't check the separate callbackBindings class property.
When iterating over the JSON fields check for registered CallbackBinding, so it doesn't execute handleRaw() unnecessarily before the CallbackBinding is handled below.